% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-pl.R
\name{assign_pl}
\alias{assign_pl}
\title{Use a Machine Learning Model to Predict Profiles' Performance Levels}
\usage{
assign_pl(
  fitted_model,
  ratings_data,
  att_levels,
  num_pls,
  rating_id = "rating",
  output_dir
)
}
\arguments{
\item{fitted_model}{An object with the fitted machine learning model.}

\item{ratings_data}{A tibble with the profiles rated by the panelists and
the panelists' ratings in long format.}

\item{att_levels}{A numeric value for the number of levels where mastery can
be demonstrated. For example, \code{att_level} is 1 for a dichotomous attribute
(i.e., nonmastery or mastery), and \code{att_level} is 2 for attributes where the
possible scores are 0, 1, and 2.}

\item{num_pls}{The number of performance levels that can be assigned to any
profile.}

\item{rating_id}{A character string for the field name of the panelists'
ratings (default is 'rating').}

\item{output_dir}{The directory path for saving the output.}
}
\value{
A list containing the fitted model, the model predictions to the
rated profiles, and the predictions for all possible profiles.
}
\description{
Predicting panelists' profile ratings with a machine learning model.
}
