% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-validation-profiles.R
\name{assign_validation_profiles}
\alias{assign_validation_profiles}
\title{Title}
\usage{
assign_validation_profiles(
  panelist_configuration = NULL,
  group_configuration = NULL,
  certainty_assignments = list(very_certain = 1L, fairly_certain = 2L, fairly_uncertain =
    3L),
  certainty_thresholds = list(very_certain = 0.2, fairly_uncertain = 0.8),
  observed,
  observed_count_label = "n",
  included_total_levels_mastered,
  att_levels,
  num_pls,
  previously_rated,
  output_dir
)
}
\arguments{
\item{panelist_configuration}{A list containing the parameters for
configuring a panelist design. The allowable parameters are \code{num_panelists}
indicating the number of panelists and
\code{shared_profiles_within_certainty_across_panelists} indicating the number of
profiles that are common to all panelists. Only one of
\code{panelist_configuration} or \code{group_configuration} should be specified.}

\item{group_configuration}{A list containing parameters for configuring a
group design. The allowable parameters are \code{num_assignment_groups} indicating
the number of groups of panelists, \code{panelists_per_group} indicating
the number of panelists in each group, and
\code{shared_profiles_within_certainty_across_groups} indicating the number
of profiles within each certainty classification that are shared by panelists
across groups. Only one of \code{panelist_configuration} or \code{group_configuration}
should be specified.}

\item{certainty_assignments}{A list of integers defining the number of
profiles to assign for each performance level and each of the certainty
classifications (very_certain, fairly_certain, fairly_uncertain). The
default values are \code{very_certain = 1L}, \code{fairly_certain = 2L}, and
\code{fairly_uncertain = 3L}.}

\item{certainty_thresholds}{A list of proportions defining the threshold
between very certain and fairly certain (\code{very_certain}; default value of
.20) and the threshold between fairly certain and fairly uncertain
(\code{fairly_uncertain}; default value is .80). The certainty metric is
quantified by dividing the probability of the second most likely performance
level classification by the probability of the most likely performance level
classification. This means the default threshold of .20 separating very
certain from fairly certain reflects profiles where the most likely
performance level is at least five times the probability of the second most
likely performance level. Similarly, the default threshold of .80 separating
fairly certain from fairly uncertain reflects profiles where the most likely
performance level is no more than 1.25 times the probability of the second
most likely performance level.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in \code{observed} (default is 'n').}

\item{included_total_levels_mastered}{The increment of the total levels
mastered, based on the attribute mastery profiles, that should be retained
during profile assignment.}

\item{att_levels}{A numeric value for the number of levels where mastery can
be demonstrated. For example, \code{att_level} is 1 for a dichotomous attribute
(i.e., nonmastery or mastery), and \code{att_level} is 2 for attributes where the
possible scores are 0, 1, and 2.}

\item{num_pls}{The number of performance levels that can be assigned to any
profile.}

\item{previously_rated}{A tibble with the profiles that have already been
assigned to panelists, where there is one row for each assigned profile.}

\item{output_dir}{The output directory for the profile assignments.}

\item{fitted_model}{An object with the fitted machine learning model.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Title
}
