% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_model}
\alias{fit_model}
\title{Fit Standard Setting Logistic Regression Model}
\usage{
fit_model(
  dat,
  iter = 4000,
  warmup = 2000,
  cores = 4,
  chains = 4,
  refresh = 0,
  adapt_delta = 0.95,
  max_treedepth = 15
)
}
\arguments{
\item{dat}{A data frame with 1 row per profile per rating. There are two
required columns:
\enumerate{
\item \code{y}: A 0/1 indicator representing whether that profile was rated in the
performance level of interest or higher. That is, if calculating the cut
point between levels 1 and 2, was the profile rated in level 2, 3, or 4?
\item \code{atts_mastered}: The total number of attributes mastered in the
profile that was rated.
}}

\item{iter}{The number of iterations in the posterior distribution
(default = 4,000).}

\item{warmup}{The number of warm-up iterations (default = 2,000).}

\item{cores}{The number of cores (default = 4).}

\item{chains}{The number of chains (default = 4).}

\item{refresh}{The interval between refreshing the visual display during
model estimation (default = 0, no refresh).}

\item{adapt_delta}{The likelihood of accepting the next step in traversing
the posterior distribution, which is related to step size (default = .95).}

\item{max_treedepth}{The maximum depth before making a U-turn when traversing
the posterior distribution (default = 15).}
}
\value{
A tibble with 1 row and 2 columns: \code{intercept} and \code{slope}. The one
row contains objects of type \link[posterior:draws_rvars]{posterior::draws_rvars}, which represent the
posterior distribution for each parameter.
}
\description{
Estimate a logistic regression to predict the performance level from the
total number of linkage levels mastered in each profile.
}
\details{
This function currently uses \code{\link[brms:brm]{brms::brm()}} to estimate the logistic
regression, but any logistic function would do (e.g., \code{\link[stats:glm]{stats::glm()}}). Note
that change the estimation engine would require down-stream adjustments as
well, as the post-processing and predictions currently assume that posterior
distributions will need to be summarized.
}
