% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_value}
\alias{max_value}
\title{Ensure a maximum value}
\usage{
max_value(x, max_val = 0.9999)
}
\arguments{
\item{x}{A double that should be capped at a specified value.}

\item{max_val}{The maximum value that \code{x} should be allowed to take.}
}
\value{
A double no larger than \code{max_val}.
}
\description{
This is a utility function to ensure functions execute correctly. For
example, when calculating Z-score, a cumulative percent of 1.0 is undefined.
This function can be used to replace all of the 1.0 with some other value,
(e.g., 0.9999) to ensure that a Z-score is calculated for each score point.
}
