% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profiles_to_retain}
\alias{profiles_to_retain}
\title{Retaining Profiles From Rangefinding}
\usage{
profiles_to_retain(pinpoint_ranges, max_levels_mastered)
}
\arguments{
\item{pinpoint_ranges}{A tibble containing the output of \code{fit_lr()}.}

\item{max_levels_mastered}{An integer specifying the maximum number of levels
that can be mastered across all attributes.}
}
\value{
A vector containing the levels of the number of skills mastered to
retain based on the output of \code{fit_lr()}.
}
\description{
Define a function to extract the levels of the total number of attributes
mastered to retain based on the logistic regression models fit using
\code{fit_lr()}.
}
