% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refine_eligible_profiles}
\alias{refine_eligible_profiles}
\title{Refine Profiles to Minimize Similarity of the Assigned Profiles}
\usage{
refine_eligible_profiles(
  profiles,
  filter_function = "median",
  filter_percentile = NULL,
  raters,
  profiles_per_level
)
}
\arguments{
\item{profiles}{A tibble containing the profiles that are eligible to be
assigned to panelists. This tibble should have fields for each of the
attributes and one field (\code{total}) for the total number of attributes
mastered.}

\item{filter_function}{The character string of the function to use for
filtering profiles. The supported options are \code{median} and \code{mean}.}

\item{filter_percentile}{A numeric value ranging from 0 to 1 to indicate the
percentile used to filter out similar profiles (default is \code{NULL}). For
example, a value of .60 indicates profiles with a Hamming distance below the
60th percentile will be filtered out of the set of eligible profiles.}

\item{raters}{A character vector containing the rater names.}

\item{profiles_per_level}{An integer value indicating the number of profiles
to sample from each level of the number of attributes mastered.}
}
\value{
A tibble with the eligible profiles and the Hamming distance.
}
\description{
This is a utility function to filter out profiles that are most similar to
the most recently assigned profile.
}
