% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slice_stratified}
\alias{slice_stratified}
\title{Stratified Sampling}
\usage{
slice_stratified(x, by, size, weight_by = NULL)
}
\arguments{
\item{x}{A tibble with one row for each rater's rating of an assigned
profile.}

\item{by}{A character field that defines the stratifications.}

\item{size}{A character field that defines the sampling sizes within each
stratification.}

\item{weight_by}{A character field that defines the sampling weights within
each stratification.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the pinpoint
ranges to assign profiles during the second round of standard setting.
}
\description{
Define a function for stratified sampling that allows for specifying the
number of samples to take within each stratification while weighting the
probability of sampling each case.
}
