% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-profiles.R
\name{assign_profiles}
\alias{assign_profiles}
\title{Title}
\usage{
assign_profiles(
  panelist_configuration = NULL,
  group_configuration = NULL,
  observed,
  observed_count_label = "n",
  included_total_levels_mastered,
  profiles_per_total_levels_mastered,
  previously_rated = NULL,
  output_dir
)
}
\arguments{
\item{panelist_configuration}{A list containing the parameters for
configuring a panelist design. The allowable parameters are \code{num_panelists}
indicating the number of panelists and \code{shared_across}
indicating the number of profiles that are common to all panelists. Only one
of \code{panelist_configuration} or \code{group_configuration} should be specified.}

\item{group_configuration}{A list containing parameters for configuring a
group design. The allowable parameters are \code{num_assignment_groups} indicating
the number of groups of panelists, \code{panelists_per_group} indicating
the number of panelists in each group, \code{shared_within}
indicating the number of profiles that are common to all of the panelists
within each group, and \code{shared_across} indicating the number
of profiles that are shared by panelists across groups. Only one of
\code{panelist_configuration} or \code{group_configuration} should be specified.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in \code{observed} (default is 'n').}

\item{included_total_levels_mastered}{The increment of the total levels
mastered, based on the attribute mastery profiles, that should be retained
during profile assignment.}

\item{profiles_per_total_levels_mastered}{An integer specifying the number of
profiles to assign to each rater at each total number of levels mastered.}

\item{previously_rated}{A tibble with the profiles that have already been
assigned to panelists, where there is one row for each assigned profile.}

\item{output_dir}{The output directory for the profile assignments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Title
}
