% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-profiles.R
\name{assign_profiles}
\alias{assign_profiles}
\title{Title}
\usage{
assign_profiles(
  num_assignment_groups,
  table_configuration = NULL,
  possible_profiles,
  observed,
  observed_count_label = "n",
  range_of_profiles,
  profiles_per_level,
  output_dir
)
}
\arguments{
\item{num_assignment_groups}{The number of "assignment groups" for the
profile assignments. When \code{table_design} is FALSE, the number of assignment
groups is the number of panelists. When \code{table_design} is TRUE, the number of
assignment groups is the number of tables of panelists.}

\item{table_configuration}{A list containing parameters for configuring a
table design. The allowable parameters are \code{panelists_per_table} indicating
the number of panelists at each table and \code{proportion_of_shared_profiles}
indicating the proportion of profiles that are common to all of the panelists
at each table.}

\item{possible_profiles}{A tibble with one row for each attribute mastery
profiles that is possible to be assigned to raters.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in \code{observed} (default is 'n').}

\item{range_of_profiles}{The increment of the total skills mastered, based on
the attribute mastery profiles, that should be retained during profile
assignment.}

\item{profiles_per_level}{An integer specifying the number of profiles to
assign to each rater at each level of the total skills mastered.}

\item{output_dir}{The output directory for the profile assignments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Title
}
