% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-profiles.R
\name{assign_profiles}
\alias{assign_profiles}
\title{Title}
\usage{
assign_profiles(
  num_assignment_groups,
  table_configuration = NULL,
  observed,
  observed_count_label = "n",
  observed_proportion_label = "prop",
  shared_profiles = 1L,
  included_totals,
  profiles_per_level,
  assigned_profiles = NULL,
  output_dir
)
}
\arguments{
\item{num_assignment_groups}{The number of "assignment groups" for the
profile assignments. When \code{table_design} is FALSE, the number of assignment
groups is the number of panelists. When \code{table_design} is TRUE, the number of
assignment groups is the number of tables of panelists.}

\item{table_configuration}{A list containing parameters for configuring a
table design. The allowable parameters are \code{panelists_per_table} indicating
the number of panelists at each table and \code{proportion_of_shared_profiles}
indicating the proportion of profiles that are common to all of the panelists
at each table.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in \code{observed} (default is 'n').}

\item{observed_proportion_label}{A character string for the field name of the
observed proportions in \code{observed} (default is 'prop').}

\item{shared_profiles}{The number of profiles that are shared by all
panelists (default is 1). With a table-based design, this is the number of
profiles that are seen by all tables. With a panelist-based design, this is
the number of profiles seen by all panelists.}

\item{included_totals}{The increment of the total skills mastered, based on
the attribute mastery profiles, that should be retained during profile
assignment.}

\item{profiles_per_level}{An integer specifying the number of profiles to
assign to each rater at each level of the total skills mastered.}

\item{assigned_profiles}{A tibble with the profiles that have already been
assigned to panelists, where there is one row for each assigned profile.}

\item{output_dir}{The output directory for the profile assignments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Title
}
