% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-profiles.R
\name{assign_profiles}
\alias{assign_profiles}
\title{Title}
\usage{
assign_profiles(
  num_assignment_groups,
  table_design = FALSE,
  eligible_profiles,
  observed,
  observed_id = "n",
  profiles_per_rater,
  raters_per_profile,
  profiles_seen_by_all = 1L,
  round,
  num_pls,
  range_of_profiles = NULL,
  output_dir
)
}
\arguments{
\item{num_assignment_groups}{The number of "assignment groups" for the
profile assignments. When \code{table_design} is FALSE, the number of assignment
groups is the number of panelists. When \code{table_design} is TRUE, the number of
assignment groups is the number of tables of panelists.}

\item{table_design}{Logical value indicating whether the panelists are
grouped into tables and profile assigned is performed at the table level.
The default value is FALSE.}

\item{eligible_profiles}{A tibble with one row for each attribute mastery
profiles that is eligible for assignment to raters.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_id}{A character string for the field name of the observed
sample sizes in the observed parameter (default is 'n').}

\item{profiles_per_rater}{A numeric value indicating the number of profiles
to be rated by each rater.}

\item{raters_per_profile}{The minimum number of raters who need to rate each
#' profile.}

\item{profiles_seen_by_all}{The number of profiles that should be rated by
all raters (default is 1).}

\item{round}{The round number of the standard setting event.}

\item{num_pls}{The number of performance levels that can be assigned to any
profile.}

\item{range_of_profiles}{The increment of the total skills mastered, based on
the attribute mastery profiles, that should be retained for profile
assignment. This should only be provided for Round 1.}

\item{output_dir}{The output directory for the profile assignments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Title
}
