% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_hamming}
\alias{calculate_hamming}
\title{Calculate Hamming distance}
\usage{
calculate_hamming(profiles, assigned_profiles, att_vec)
}
\arguments{
\item{profiles}{A tibble containing the profiles that are eligible to be
assigned to panelists. This tibble should have fields for each of the
attributes and one field (\code{total}) for the total number of attributes
mastered.}

\item{assigned_profiles}{A tibble containing the assigned profiles. This
tibble should only contain fields for each of the attribute.}

\item{att_vec}{A character vector containing the attribute names.}
}
\value{
A tibble with the eligible profiles and the Hamming distance.
}
\description{
This is a utility function to calculate the Hamming distance between the last
assigned profile and the remaining profiles that are eligible for assignment.
}
