% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-agreement.R
\name{eval_agreement}
\alias{eval_agreement}
\title{Evaluating the Machine Learning Model}
\usage{
eval_agreement(
  model_ratings,
  metrics,
  observed,
  num_pls,
  rating_id = "rating",
  observed_count_label = "n",
  observed_proportion_label = "prop",
  output_dir
)
}
\arguments{
\item{model_ratings}{A tibble containing the assigned profiles along with
the model-predicted performance level, the model-predicted probability for
each performance level, and the rating assigned by each panelist. This tibble
is the output of \code{predict_ml()} and is one row per panelist rating.}

\item{metrics}{A character vector containing the evaluation metrics that
should be included in the output. Can include any of \code{"accuracy"},
\code{"adjacent"}, \code{"kappa"}, \code{"auc"}, \code{"gwet"}, or \code{"assignment"}. Including
\code{"accuracy"} calculates classification accuracy. Including \code{"adjacent"}
calculates adjacent classification accuracy. Including \code{"kappa"} calculates
Cohen's kappa. Including \code{"auc"} calculates the area under the receiver
operating characteristic curve. Including \code{"gwet"} calculates Gwet's $AC_2$.
Including \code{"assignment"} calculates assignment statistics from the standard
setting procedure -- the number of profiles assigned, the number of students
with the assigned profiles, and the proportion of students with the assigned
profiles.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{num_pls}{The number of performance levels that can be assigned to any
profile.}

\item{rating_id}{A character string for the field name of the panelists'
ratings (default is 'rating').}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in \code{observed} (default is 'n').}

\item{observed_proportion_label}{A character string for the field name of the
observed proportions in \code{observed} (default is 'prop').}

\item{output_dir}{The directory path for saving the output.}
}
\value{
A list containing the fitted model, the model predictions to the
rated profiles, and the predictions for all possible profiles.
}
\description{
Evaluating the machine learning model.
}
