% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lr.R
\name{fit_lr}
\alias{fit_lr}
\title{Range Finding Procedure from the Condensed Mastery Standard Setting Method}
\usage{
fit_lr(
  ratings,
  meta_data,
  pl_labels,
  max_score,
  cores = 4,
  chains = 4,
  output_dir
)
}
\arguments{
\item{ratings}{A tibble with one row for each rater's rating of an assigned
profile, including columns for the profile number, the mastery status for
each attribute, and the raters' ratings.}

\item{meta_data}{A tibble with the profile meta-data, including the profile
id and columns for each attribute and the number of levels mastered for
each attribute.}

\item{pl_labels}{A character vector containing the ordered performance
levels.}

\item{max_score}{A numeric value for the maximum score for the number of
skills that can be mastered within an attribute. For example, \code{max_score} is
1 for a dichotomous attribute (i.e., nonmastery or mastery), and \code{max_score}
is 2 for attributes where the possible scores are 0, 1, and 2.}

\item{cores}{The number of cores (default = 4).}

\item{chains}{The number of chains (default = 4).}

\item{output_dir}{The output directory for the pinpointing ranges.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the pinpoint
ranges to assign profiles during the second round of standard setting.
}
\description{
Fit a logistic regression model to conduct the range finding procedure from
the condensed mastery standard setting method.
}
