% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lr.R
\name{fit_lr}
\alias{fit_lr}
\title{Range Finding Procedure from the Condensed Mastery Standard Setting Method}
\usage{
fit_lr(ratings, pl_labels, att_levels, cores = 4, chains = 4, output_dir)
}
\arguments{
\item{ratings}{A tibble with one row for each rater's rating of an assigned
profile, including columns for the profile number, the mastery status for
each attribute, and the raters' ratings.}

\item{pl_labels}{A character vector containing the ordered performance
levels.}

\item{att_levels}{An integer describing the number of categorical mastery
classes for each attribute.}

\item{cores}{The number of cores (default = 4).}

\item{chains}{The number of chains (default = 4).}

\item{output_dir}{The output directory for the pinpointing ranges.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the pinpoint
ranges to assign profiles during the second round of standard setting.
}
\description{
Fit a logistic regression model to conduct the range finding procedure from
the condensed mastery standard setting method.
}
