% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ml.R
\name{fit_ml}
\alias{fit_ml}
\title{Fitting the Machine Learning Model}
\usage{
fit_ml(
  user_workflow,
  ratings_data,
  possible_profiles,
  observed,
  att_levels,
  num_pls,
  output_dir
)
}
\arguments{
\item{user_workflow}{A tibble with one row for each attribute mastery
profiles that is eligible for assignment to raters.}

\item{ratings_data}{A tibble with the profiles rated by the panelists and
the panelists' ratings in long format.}

\item{possible_profiles}{A tibble with all of the possible profiles.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{att_levels}{A numeric value for the number of levels where mastery can
be demonstrated. For example, \code{att_level} is 1 for a dichotomous attribute
(i.e., nonmastery or mastery), and \code{att_level} is 2 for attributes where the
possible scores are 0, 1, and 2.}

\item{num_pls}{The number of performance levels that can be assigned to any
profile.}

\item{output_dir}{The directory path for saving the output.}
}
\value{
A list containing the fitted model, the model predictions to the
rated profiles, and the predictions for all possible profiles.
}
\description{
Fit a machine learning model to the panelists' profile ratings.
}
