% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{pmp-class}
\alias{pmp-class}
\alias{pmp}
\title{S7 class for model estimation and evaluation}
\usage{
pmp(
  model = .no_constructor(),
  in_sample_agreement = .no_constructor(),
  out_of_sample_agreement = .no_constructor()
)
}
\arguments{
\item{model}{The workflow for the fitted model.}

\item{in_sample_agreement}{A tibble containing the in-sample agreement
statistics.}

\item{out_of_sample_agreement}{A tibble containing the out-of-sample
agreement statistics.}
}
\value{
An \link[S7:S7_object]{S7 object} with the corresponding class.
}
\description{
The estimated model, in-sample agreement metrics, and out-of-sample
agreement metrics are exported as an S7 class.
}
