% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premp-package.R
\docType{package}
\name{premp-package}
\alias{premp}
\alias{premp-package}
\title{premp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-dcm/premp}
  \item Report bugs at \url{https://github.com/r-dcm/premp/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

}
\keyword{internal}
