% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profiles_to_retain}
\alias{profiles_to_retain}
\title{Retaining Profiles From Rangefinding}
\usage{
profiles_to_retain(pinpoint_ranges, num_attributes, attribute_levels)
}
\arguments{
\item{pinpoint_ranges}{A tibble containing the output of \code{fit_lr()}.}

\item{num_attributes}{An integer specifying the number of measured
attributes.}

\item{attribute_levels}{An integer specifying the number of levels per
attribute.}
}
\value{
A vector containing the levels of the number of skills mastered to
retain based on the output of \code{fit_lr()}.
}
\description{
Define a function to extract the levels of the total number of attributes
mastered to retain based on the logistic regression models fit using
\code{fit_lr()}.
}
