% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-sampling.R
\name{weighted_sampling}
\alias{weighted_sampling}
\title{Weighted Sampling}
\usage{
weighted_sampling(
  possible_profiles,
  observed,
  observed_count_label,
  profiles_per_total_levels_mastered,
  raters,
  panelist_configuration = NULL,
  group_configuration = NULL
)
}
\arguments{
\item{possible_profiles}{A tibble with one row for each attribute mastery
profiles that is eligible for assignment to raters.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in the observed parameter.}

\item{profiles_per_total_levels_mastered}{An integer specifying the number of
profiles to assign to each rater at each total number of levels mastered.}

\item{raters}{A character vector containing the raters' ids.}

\item{panelist_configuration}{A list containing the parameters for
configuring a panelist design. The allowable parameters are \code{num_panelists}
indicating the number of panelists and \code{shared_across}
indicating the number of profiles that are common to all panelists. Only one
of \code{panelist_configuration} or \code{group_configuration} should be specified.}

\item{group_configuration}{A list containing parameters for configuring a
group design. The allowable parameters are \code{num_assignment_groups} indicating
the number of groups of panelists, \code{panelists_per_group} indicating
the number of panelists in each group, \code{shared_within}
indicating the number of profiles that are common to all of the panelists
within each group, and \code{shared_across} indicating the number
of profiles that are shared by panelists across groups. Only one of
\code{panelist_configuration} or \code{group_configuration} should be specified.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the profiles to
be assigned to raters during a standard setting event.
}
\description{
Apply a weighted sampling procedure to sample profiles that will be assigned
to raters during a sample setting event.
}
