% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-sampling.R
\name{weighted_sampling}
\alias{weighted_sampling}
\title{Weighted Sampling}
\usage{
weighted_sampling(
  possible_profiles,
  observed,
  observed_count_label,
  shared_profiles_across_groups,
  profiles_per_level,
  raters,
  group_configuration = NULL
)
}
\arguments{
\item{possible_profiles}{A tibble with one row for each attribute mastery
profiles that is eligible for assignment to raters.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in the observed parameter.}

\item{shared_profiles_across_groups}{The number of profiles that are shared
by panelists across groups. With a group-based design, this is the number of
profiles that are seen by all groups. With a panelist-based design, this is
the number of profiles seen by all panelists.}

\item{profiles_per_level}{An integer specifying the number of profiles to
assign to each rater at each level of the total skills mastered.}

\item{raters}{A character vector containing the raters' ids.}

\item{group_configuration}{A list containing parameters for configuring a
group design. The allowable parameters are \code{panelists_per_group} indicating
the number of panelists in each group and \code{shared_profiles_within_groups}
indicating the number of profiles that are common to all of the panelists
within each group.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the profiles to
be assigned to raters during a standard setting event.
}
\description{
Apply a weighted sampling procedure to sample profiles that will be assigned
to raters during a sample setting event.
}
