% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-sampling.R
\name{weighted_sampling}
\alias{weighted_sampling}
\title{Weighted Sampling}
\usage{
weighted_sampling(
  possible_profiles,
  observed,
  observed_count_label,
  observed_proportion_label,
  shared_profiles,
  profiles_per_level,
  raters,
  table_configuration = NULL
)
}
\arguments{
\item{possible_profiles}{A tibble with one row for each attribute mastery
profiles that is eligible for assignment to raters.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{observed_count_label}{A character string for the field name of the
observed sample sizes in the observed parameter.}

\item{observed_proportion_label}{A character string for the field name of the
observed proportions in the observed parameter.}

\item{shared_profiles}{The number of profiles that are shared by all
panelists. With a table-based design, this is the number of profiles that are
seen by all tables. With a panelist-based design, this is the number of
profiles seen by all panelists.}

\item{profiles_per_level}{An integer specifying the number of profiles to
assign to each rater at each level of the total skills mastered.}

\item{raters}{A character vector containing the raters' ids.}

\item{table_configuration}{A list containing parameters for configuring a
table design. The allowable parameters are \code{panelists_per_table} indicating
the number of panelists at each table and \code{proportion_of_shared_profiles}
indicating the proportion of profiles that are common to all of the panelists
at each table.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the profiles to
be assigned to raters during a standard setting event.
}
\description{
Apply a weighted sampling procedure to sample profiles that will be assigned
to raters during a sample setting event.
}
