% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-sampling.R
\name{weighted_sampling}
\alias{weighted_sampling}
\title{Weighted Sampling}
\usage{
weighted_sampling(eligible_profiles, num_profiles, observed, round, num_pls)
}
\arguments{
\item{eligible_profiles}{A tibble with one row for each attribute mastery
profiles that is eligible for assignment to raters.}

\item{num_profiles}{The maximum number of profiles that be assigned during
this round of the standard setting event.}

\item{observed}{A tibble with one row for each attribute mastery profile that
was observed along with the number of times it was observed.}

\item{round}{The round number of the standard setting event.}

\item{num_pls}{The number of performance levels that can be assigned to any
profile.}
}
\value{
\link[tibble:tibble-package]{tibble} A tibble containing the profiles to
be assigned to raters during a standard setting event.
}
\description{
Apply a weighted sampling procedure to sample profiles that will be assigned
to raters during a sample setting event.
}
